Rename sol_reader -> asl_sol_reader
